requirements.txt
vertexai
python-dotenv
rich
pickle
